Upgraded underscore (1.12.1)
The package underscore from 1.13.0-0 and before 1.13.0-2, from 1.3.2 and before 1.12.1 are vulnerable to Arbitrary Code Execution via the template function, particularly when a variable property is passed as an argument as it is not sanitized.
Updating dependency to minimum fix version (1.12.1). This in an attempt to fix the vulnerability with minimal chance of breaking the code by using the newest version.